Refactor: Consolidate UI components and update core classes
- Moved UI components from ui/ to toolbox/+bctui/ for better package organization
- Updated Manifold, Signal, and Graph classes with improved functionality
- Removed legacy UI structure and documentation from ui/ directory
- Added new brush system, kernel factories, and component architecture
- Added comprehensive examples and test files
- Updated bioctree_start.m with improved initialization
- Added new experimental features for DEC and streamline analysis
- Updated .gitignore to exclude large data files and debug logs

+++ toolbox/+bct/@Manifold/Manifold.m
@@ -212,7 +212,7 @@ classdef Manifold < bct.Domain
             % See also: triangulation
             
             if isempty(obj.Triangulation)
-                obj.Triangulation = triangulation(obj.Faces, obj.Vertices);
+                obj.Triangulation = triangulation(double(obj.Faces), double(obj.Vertices));
             end
             TR = obj.Triangulation;
         end
